+++ specs/__snapshot__/demo-0.js
@@ -3,4 +3,4 @@
 const React = __UMDBindCheck(globalThis["React"], [], "react", "globalThis.React", false);
 import "./polyfill.js";
 console.log(React);
-import { __UMDBindCheck as __UMDBindCheck } from "https://cdn.jsdelivr.net/npm/@magic-works/ttypescript-browser-like-import-transformer@2.0.2/es/ttsclib.min.js";
+import { __UMDBindCheck as __UMDBindCheck } from "https://cdn.jsdelivr.net/npm/@magic-works/ttypescript-browser-like-import-transformer@2.0.5/es/ttsclib.min.js";