+++ specs/__snapshot__/globalObject-undefined.js
@@ -2,4 +2,4 @@
 // PluginConfig: {}
 const R = __UMDBindCheck(globalThis["React"], ["default"], "react", "globalThis.React", false).default;
 console.log(R);
-import { __UMDBindCheck as __UMDBindCheck } from "https://cdn.jsdelivr.net/npm/@magic-works/ttypescript-browser-like-import-transformer@2.0.2/es/ttsclib.min.js";
+import { __UMDBindCheck as __UMDBindCheck } from "https://cdn.jsdelivr.net/npm/@magic-works/ttypescript-browser-like-import-transformer@2.0.5/es/ttsclib.min.js";