+++ src/core.ts
@@ -1,4 +1,4 @@
-import {
+import type {
     TransformationContext,
     SourceFile,
     Node,
@@ -21,15 +21,14 @@ import {
     StringLiteral,
     NumericLiteral,
     ArrayLiteralExpression,
-    FunctionDeclaration,
     BooleanLiteral,
     Program,
     PropertyAccessExpression,
     NamedImports,
     CompilerOptions,
 } from 'typescript'
-import { PluginConfigs, ImportMapFunctionOpts, BareModuleRewriteUMD } from './plugin-config'
-import { NormalizedPluginConfig } from './config-parser'
+import type { PluginConfigs, ImportMapFunctionOpts, BareModuleRewriteUMD } from './plugin-config'
+import type { NormalizedPluginConfig } from './config-parser'
 type ts = typeof import('typescript')
 
 export interface CustomTransformationContext<T extends Node> {
@@ -75,7 +74,7 @@ export default function createTransformer(
 ) {
     const { ts, configParser, importMapResolve } = core
     // ? Don't rely on the ts.Program because don't want to create on during the test.
-    return function(_program: Partial<Pick<Program, 'getCurrentDirectory'>>, configRaw: PluginConfigs) {
+    return function (_program: Partial<Pick<Program, 'getCurrentDirectory'>>, configRaw: PluginConfigs) {
         return (context: TransformationContext) => {
             configParser.validateConfig(configRaw, context.getCompilerOptions())
             const config = configParser.normalizePluginConfig(configRaw)
@@ -107,7 +106,9 @@ export default function createTransformer(
 
                 let visitedSourceFile = ts.visitEachChild(sourceFile, visitor, context)
                 // ? hoistedHelper and hoistedUMDImport will be added ^ in the visitor
-                const hoistedHelper = Array.from(topLevelScopedHelperMap.get(sourceFile)?.values() || []).map(x => x[1])
+                const hoistedHelper = Array.from(topLevelScopedHelperMap.get(sourceFile)?.values() || []).map(
+                    (x) => x[1],
+                )
                 const ttscHelper = ttsclibImportMap.get(sourceFile)
                 if (ttscHelper) hoistedHelper.push(ttscHelper[0])
                 function isLanguageHoistable(node: Statement): boolean {
@@ -116,7 +117,7 @@ export default function createTransformer(
                     return false
                 }
                 const languageHoistableDeclarations = hoistedHelper.filter(isLanguageHoistable)
-                const languageNotHoistableDeclarations = hoistedHelper.filter(x => !isLanguageHoistable(x))
+                const languageNotHoistableDeclarations = hoistedHelper.filter((x) => !isLanguageHoistable(x))
                 const hoistedUMDImport = Array.from(hoistUMDImportDeclaration.get(sourceFile)?.values() || [])
                 visitedSourceFile = ts.updateSourceFileNode(
                     visitedSourceFile,
@@ -151,7 +152,6 @@ export default function createTransformer(
                     const shared = {
                         config,
                         configRaw,
-                        ts,
                         context,
                         node,
                         sourceFile,
@@ -215,14 +215,14 @@ function getImportedItems(ts: ts, node: ImportDeclaration | ExportDeclaration): 
             const b = clause.namedBindings
             if (ts.isNamespaceImport(b)) result.add('*')
             else if (ts.isNamedImports(b)) {
-                b.elements.forEach(x => result.add(x.name.text))
+                b.elements.forEach((x) => result.add((x.propertyName || x.name).text))
             }
         }
     } else {
         if (!node.exportClause) return new Set('*')
         const clause = node.exportClause
         if (ts.isNamedExports(clause)) {
-            clause.elements.forEach(x => result.add(x.name.text))
+            clause.elements.forEach((x) => result.add((x.propertyName || x.name).text))
         }
         if (ts?.isNamespaceExport(clause)) result.add('*')
     }
@@ -314,8 +314,8 @@ function updateImportExportDeclaration(
                 return [ts.createExpressionStatement(ts.createLiteral(text))]
             }
             const { statements } = importOrExportClauseToUMD(nextPath, _with(context, clause), globalObject)
-            writeSourceFileMeta(sourceFile, hoistUMDImportDeclaration, new Set<Statement>(), _ => {
-                statements.forEach(x => _.add(x))
+            writeSourceFileMeta(sourceFile, hoistUMDImportDeclaration, new Set<Statement>(), (_) => {
+                statements.forEach((x) => _.add(x))
             })
             return [...umdImport, ...statements]
         }
@@ -355,7 +355,7 @@ function importOrExportClauseToUMD(
             statements.push(getNamespaceImport(nsImport))
         }
         if (namedImport) {
-            namedImport.elements.forEach(v => ids.push(v.name))
+            namedImport.elements.forEach((v) => ids.push(v.name))
             statements.push(transformNamedImportExport(namedImport))
         }
         return { variableNames: ids, statements: statements }
@@ -367,7 +367,7 @@ function importOrExportClauseToUMD(
             ts.createImportClause(
                 undefined,
                 ts.createNamedImports(
-                    node.elements.map<ImportSpecifier>(x => {
+                    node.elements.map<ImportSpecifier>((x) => {
                         const id = ts.createFileLevelUniqueName(x.name.text)
                         ghostBindings.set(x, id)
                         return ts.createImportSpecifier(x.propertyName, id)
@@ -416,16 +416,16 @@ function importOrExportClauseToUMD(
         // ? If the UMD binding is undefined, this will give a warning but success (with an empty module).
         return getAssignment(
             namedImport.name,
-            createCheckedUMDAccess(x => x),
+            createCheckedUMDAccess((x) => x),
         )
     }
     function getDefaultImport(defaultImport: Identifier) {
         let umdAccess: Expression
         if (esModuleInterop) {
-            const [, createESModuleInteropCall] = createTopLevelScopedHelper(ctx, ttsclib.__esModuleInterop, [])
+            const [, createESModuleInteropCall] = createTopLevelScopedHelper(ctx, ttsclib.__esModuleInterop)
             umdAccess = createCheckedUMDAccess(createESModuleInteropCall, ts.createLiteral('default'))
         } else {
-            umdAccess = createCheckedUMDAccess(x => x, ts.createLiteral('default'))
+            umdAccess = createCheckedUMDAccess((x) => x, ts.createLiteral('default'))
         }
         // ? const defaultImportIdentifier = __importBindingCheck(value, name, path, mappedName)
         return getAssignment(defaultImport, ts.createPropertyAccess(umdAccess, 'default'))
@@ -447,12 +447,12 @@ function importOrExportClauseToUMD(
                 [
                     ts.createVariableDeclaration(
                         ts.createObjectBindingPattern(
-                            elements.map(x => ts.createBindingElement(undefined, x.propertyName, x.name, undefined)),
+                            elements.map((x) => ts.createBindingElement(undefined, x.propertyName, x.name, undefined)),
                         ),
                         undefined,
                         createCheckedUMDAccess(
-                            x => x,
-                            ...elements.map(x => ts.createLiteral(x.propertyName?.text ?? x.name.text)),
+                            (x) => x,
+                            ...elements.map((x) => ts.createLiteral(x.propertyName?.text ?? x.name.text)),
                         ),
                     ),
                 ],
@@ -461,7 +461,7 @@ function importOrExportClauseToUMD(
         )
     }
     function createCheckedUMDAccess(wrapper: (x: Expression) => Expression, ...names: StringLiteral[]) {
-        const [, createUMDBindCheck] = createTopLevelScopedHelper(ctx, ttsclib.__UMDBindCheck, [])
+        const [, createUMDBindCheck] = createTopLevelScopedHelper(ctx, ttsclib.__UMDBindCheck)
         return createUMDBindCheck(
             wrapper(umdAccess),
             ts.createArrayLiteral(names),
@@ -495,7 +495,7 @@ function getUMDExpressionForModule(
               ts,
               `${globalObject}.${umdName}`,
               'expression',
-              function(node): node is CallExpression | PropertyAccessExpression {
+              function (node): node is CallExpression | PropertyAccessExpression {
                   return ts.isCallExpression(node) || ts.isPropertyAccessExpression(node)
               },
               sourceFile.fileName,
@@ -522,7 +522,7 @@ function transformDynamicImport(ctx: Omit<Context<CallExpression>, 'path'>, args
         })
         switch (rewriteStrategy.type) {
             case 'error':
-                const [id] = createTopLevelScopedHelper(ctx, dynamicImportFailedHelper(args), [])
+                const [id] = createTopLevelScopedHelper(ctx, dynamicImportFailedHelper(args))
                 return [ts.createCall(id, void 0, [ts.createLiteral(rewriteStrategy.message), ...args])]
             case 'noop':
                 return [node]
@@ -531,7 +531,7 @@ function transformDynamicImport(ctx: Omit<Context<CallExpression>, 'path'>, args
             }
             case 'umd': {
                 if (rest.length !== 0) {
-                    const [id] = createTopLevelScopedHelper(ctx, dynamicImportFailedHelper(args), [])
+                    const [id] = createTopLevelScopedHelper(ctx, dynamicImportFailedHelper(args))
                     return [
                         ts.createCall(id, void 0, [
                             ts.createLiteral(moreThan1ArgumentDynamicImportErrorMessage),
@@ -558,7 +558,7 @@ function transformDynamicImport(ctx: Omit<Context<CallExpression>, 'path'>, args
         }
     } else {
         if (rest.length !== 0) {
-            const [id] = createTopLevelScopedHelper(ctx, dynamicImportFailedHelper(args), [])
+            const [id] = createTopLevelScopedHelper(ctx, dynamicImportFailedHelper(args))
             return [ts.createCall(id, void 0, [ts.createLiteral(moreThan1ArgumentDynamicImportErrorMessage), ...args])]
         }
         const opt = config.dynamicImportPathRewrite
@@ -572,14 +572,13 @@ function transformDynamicImport(ctx: Omit<Context<CallExpression>, 'path'>, args
         const [dynamicImportTransformIdentifier, createDynamicImportTransform] = createTopLevelScopedHelper(
             ctx,
             ttsclib.__dynamicImportTransform,
-            [moduleSpecifierTransform_.value],
         )
         const stringifiedConfig = ts.createCall(ts.createIdentifier('JSON.parse'), void 0, [
             ts.createLiteral(JSON.stringify(config)),
         ])
-        const [dynamicImportNative] = createTopLevelScopedHelper(ctx, dynamicImportNativeString, [])
-        const [umdBindCheck] = createTopLevelScopedHelper(ctx, ttsclib.__UMDBindCheck, [])
-        const [moduleSpecifierTransform] = createTopLevelScopedHelper(ctx, ttsclib.moduleSpecifierTransform, [])
+        const [dynamicImportNative] = createTopLevelScopedHelper(ctx, dynamicImportNativeString)
+        const [umdBindCheck] = createTopLevelScopedHelper(ctx, ttsclib.__UMDBindCheck)
+        const [moduleSpecifierTransform] = createTopLevelScopedHelper(ctx, ttsclib.moduleSpecifierTransform)
         const helperArgs = [stringifiedConfig, dynamicImportNative, umdBindCheck, moduleSpecifierTransform] as const
         if (opt === 'auto' || opt === undefined) {
             /**
@@ -603,11 +602,16 @@ function transformDynamicImport(ctx: Omit<Context<CallExpression>, 'path'>, args
                     ts.NodeFlags.Const,
                 ),
             )
-            writeSourceFileMeta(sourceFile, topLevelScopedHelperMap, new Map<string, [Identifier, Statement]>(), x => {
-                x.set('__customImportHelper', [customFunction, decl])
-            })
+            writeSourceFileMeta(
+                sourceFile,
+                topLevelScopedHelperMap,
+                new Map<string, [Identifier, Statement]>(),
+                (x) => {
+                    x.set('__customImportHelper', [customFunction, decl])
+                },
+            )
         }
-        const [, createCustomImportHelper] = createTopLevelScopedHelper(ctx, ttsclib.__customDynamicImportHelper, [])
+        const [, createCustomImportHelper] = createTopLevelScopedHelper(ctx, ttsclib.__customDynamicImportHelper)
         /**
          * __customImportHelper(
          *     path,
@@ -648,7 +652,6 @@ const ttsclibImportMap = new Map<SourceFile, [ImportDeclaration, Set<Identifier>
 function createTopLevelScopedHelper<F extends string | ((...args: any[]) => any)>(
     context: Pick<Context<any>, 'ts' | 'sourceFile' | 'config' | 'ttsclib' | 'queryPackageVersion'>,
     helper: F | string,
-    additionDeclarations: FunctionDeclaration[],
 ): readonly [Identifier, (...args: CastArray<LevelUpArgs<Parameters<CastFunction<F>>>>) => Expression] {
     const { config, ts, sourceFile, ttsclib, queryPackageVersion } = context
     const result = topLevelScopedHelperMap.get(sourceFile)?.get(helper)
@@ -691,8 +694,8 @@ function createTopLevelScopedHelper<F extends string | ((...args: any[]) => any)
         const importBind = ts.createImportSpecifier(uniqueName, ts.createIdentifier(fnName))
         const importClause = importDec.importClause!
         const importBindings = importClause.namedBindings! as NamedImports
-        if (importBindings.elements.find(x => x.name.text === fnName)) {
-            const uniqueName = [...idSet.values()].find(x => x.text === fnName)!
+        if (importBindings.elements.find((x) => x.name.text === fnName)) {
+            const uniqueName = [...idSet.values()].find((x) => x.text === fnName)!
             return [uniqueName, (...args: any) => ts.createCall(uniqueName, void 0, args)] as const
         }
         importDec = ts.updateImportDeclaration(
@@ -720,13 +723,11 @@ function createTopLevelScopedHelper<F extends string | ((...args: any[]) => any)
         void 0,
         parsedFunction.parameters,
         void 0,
-        parsedFunction.body
-            ? ts.updateBlock(parsedFunction.body, [...parsedFunction.body.statements, ...additionDeclarations])
-            : parsedFunction.body,
+        parsedFunction.body,
     )
     ts.setEmitFlags(f, ts.EmitFlags.NoComments)
     ts.setEmitFlags(f, ts.EmitFlags.NoNestedComments)
-    writeSourceFileMeta(sourceFile, topLevelScopedHelperMap, new Map(), x => x.set(helper, [uniqueName, f]))
+    writeSourceFileMeta(sourceFile, topLevelScopedHelperMap, new Map(), (x) => x.set(helper, [uniqueName, f]))
     return returnValue
 }
 
@@ -751,7 +752,7 @@ function parseJS<T extends Node = Node>(
         }
     }
     const sf = ts.createSourceFile(fileName, source, ts.ScriptTarget.ESNext, false, ts.ScriptKind.JS)
-    if (sf.statements.filter(x => !ts.isEmptyStatement(x)).length !== 1)
+    if (sf.statements.filter((x) => !ts.isEmptyStatement(x)).length !== 1)
         return { type: 'syntax error', error: 'Unexpected statement count ' + sf.statements.length }
     const [firstStatement] = sf.statements
     if (kind === 'expression') {